refactor: 유즈 케이스 패키지 위치 변경 (#45)
