+++ src/main/kotlin/kr/wooco/woocobe/auth/domain/usecase/WithdrawUseCase.kt
@@ -3,7 +3,7 @@ package kr.wooco.woocobe.auth.domain.usecase
 import kr.wooco.woocobe.auth.domain.gateway.AuthTokenStorageGateway
 import kr.wooco.woocobe.auth.domain.gateway.AuthUserStorageGateway
 import kr.wooco.woocobe.auth.domain.gateway.TokenProviderGateway
-import kr.wooco.woocobe.common.domain.UseCase
+import kr.wooco.woocobe.common.domain.usecase.UseCase
 import kr.wooco.woocobe.user.domain.gateway.UserStorageGateway
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
