+++ src/main/kotlin/kr/wooco/woocobe/common/domain/usecase/UseCase.kt
@@ -1,4 +1,4 @@
-package kr.wooco.woocobe.common.domain
+package kr.wooco.woocobe.common.domain.usecase
 
 fun interface UseCase<I, O> {
     fun execute(input: I): O
