+++ src/main/kotlin/kr/wooco/woocobe/place/domain/usecase/GetPlaceUseCase.kt
@@ -1,6 +1,6 @@
 package kr.wooco.woocobe.place.domain.usecase
 
-import kr.wooco.woocobe.common.domain.UseCase
+import kr.wooco.woocobe.common.domain.usecase.UseCase
 import kr.wooco.woocobe.place.domain.gateway.PlaceStorageGateway
 import kr.wooco.woocobe.place.domain.model.Place
 import org.springframework.stereotype.Service
