+++ src/main/kotlin/kr/wooco/woocobe/place/domain/usecase/UpdatePlaceReviewUseCase.kt
@@ -1,6 +1,6 @@
 package kr.wooco.woocobe.place.domain.usecase
 
-import kr.wooco.woocobe.common.domain.UseCase
+import kr.wooco.woocobe.common.domain.usecase.UseCase
 import kr.wooco.woocobe.place.domain.gateway.PlaceReviewStorageGateway
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
