+++ src/main/kotlin/kr/wooco/woocobe/plan/domain/usecase/DeletePlanUseCase.kt
@@ -1,7 +1,7 @@
 package kr.wooco.woocobe.plan.domain.usecase
 
 import jakarta.transaction.Transactional
-import kr.wooco.woocobe.common.domain.UseCase
+import kr.wooco.woocobe.common.domain.usecase.UseCase
 import kr.wooco.woocobe.plan.domain.gateway.PlanStorageGateway
 import org.springframework.stereotype.Service
